// dict/api/task/types.ts
/* eslint-disable */
// tslint:disable
/**
 * MICRO SERVICES CURD OPERATIONS
 * 
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface TaskApi {
	taskcreated: (request: Api.BODYDATA | undefined) => Promise<TaskcreatedResponse>
	deletEtask: (questionid: string) => Promise<DeletEtaskResponse>
	gettask: () => Promise<GettaskResponse>
	updatetask: (questionid: string, request: Api.BODYDATA | undefined) => Promise<UpdatetaskResponse>
}

export type TaskcreatedResponse = Taskcreated201Response | Taskcreated400Response

export interface Taskcreated201Response {
	status: 201
	body: Api.ApiResponse
	headers?: never
}

export interface Taskcreated400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type DeletEtaskResponse = DeletEtask201Response | DeletEtask400Response

export interface DeletEtask201Response {
	status: 201
	body: Api.ApiResponse
	headers?: never
}

export interface DeletEtask400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type GettaskResponse = Gettask201Response | Gettask400Response

export interface Gettask201Response {
	status: 201
	body: Api.BODYDATA[]
	headers?: never
}

export interface Gettask400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}

export type UpdatetaskResponse = Updatetask201Response | Updatetask400Response

export interface Updatetask201Response {
	status: 201
	body: Api.ApiResponse
	headers?: never
}

export interface Updatetask400Response {
	status: 400
	body: Api.ApiResponse
	headers?: never
}